+++ components/center/src/useWindowSize.js
@@ -1,21 +1,20 @@
 import { useState, useEffect } from "haunted";
 
 export function useWindowSize() {
-    const [windowSize, setWindowSize] = useState({
-      width: 0,
-      height: 0,
-    })
-    useEffect(() => {
-      function handleResize() {
-        setWindowSize({
-          width: window.innerWidth,
-          height: window.innerHeight,
-        })
-      }
-      window.addEventListener('resize', handleResize)
-      handleResize()
-      return () => window.removeEventListener('resize', handleResize)
-    }, [])
-    return windowSize
-  }
-  
+  const [windowSize, setWindowSize] = useState({
+    width: 0,
+    height: 0,
+  });
+  useEffect(() => {
+    function handleResize() {
+      setWindowSize({
+        width: window.innerWidth,
+        height: window.innerHeight,
+      });
+    }
+    window.addEventListener("resize", handleResize);
+    handleResize();
+    return () => window.removeEventListener("resize", handleResize);
+  }, []);
+  return windowSize;
+}