+++ components/content-switch-fade/rollup.config.js
@@ -1,7 +1,0 @@
-import configure from "../../rollup.config";
-import { dependencies } from "./package.json";
-
-export default configure({
-  input: "./src/index.js",
-  dependencies,
-});